agent/requirements.txt
anthropic>=0.40.0
web3>=7.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
eth-account>=0.12.0
